Made a few changes and deleted some useless files

+++ IrisEncoding/Subject.cpp
@@ -6,6 +6,9 @@ Subject::Subject(std::string normPath, std::string maskPath, std::string imgName
     this->imgName = imgName;
     this->normImg = cv::imread(normPath, cv::IMREAD_GRAYSCALE);
     this->mask = cv::imread(maskPath, cv::IMREAD_GRAYSCALE);
+
+    // this->encodedImg = cv::Mat::zeros(100, 600, CV_8UC1);
+    // this->alreadyExists = false;
     if(!boost::filesystem::exists("./LBP_Codes/"+imgName+"_LBP.png"))
     {
         this->encodedImg = cv::Mat::zeros(100, 600, CV_8UC1);
@@ -27,5 +30,9 @@ Subject::~Subject()
 
 void Subject::saveTemplate()
 {
-    cv::imwrite("./LBP_Codes/"+imgName+"_LBP.png", this->encodedImg);
+    if(!(this->alreadyExists))
+    {
+        cv::imwrite("./LBP_Codes/"+imgName+"_LBP.png", this->encodedImg);
+        cv::imwrite("./Masks/"+imgName+"_NORMMASK.png", this->mask);
+    }
 }
